+++ pacman/src/utils/box.cs
@@ -2,15 +2,15 @@ namespace MainNamespace
 {
   class Box
   {
-    public int Width;
-    public int Height;
-    public int Left;
-    public int Right { get => this.Left + this.Width; }
-    public int Top;
-    public int Bottom { get => this.Top + this.Height; }
-    public int YCenter { get => this.Top + this.Height / 2; }
-    public int XCenter { get => this.Left + this.Width / 2; }
-    public Box(int width, int height, int left, int top)
+    public double Width;
+    public double Height;
+    public double Left;
+    public double Right { get => this.Left + this.Width; }
+    public double Top;
+    public double Bottom { get => this.Top + this.Height; }
+    public double YCenter { get => this.Top + this.Height / 2; }
+    public double XCenter { get => this.Left + this.Width / 2; }
+    public Box(double width, double height, double left, double top)
     {
       this.Width = width;
       this.Height = height;
